Move `ResponsibleEntity` model to its own file
