+++ city_furniture/admin/common.py
@@ -1,12 +1,6 @@
-from django import forms
 from django.contrib.gis import admin
 
-from city_furniture.models.common import (
-    CityFurnitureColor,
-    CityFurnitureDeviceType,
-    CityFurnitureTarget,
-    ResponsibleEntity,
-)
+from city_furniture.models.common import CityFurnitureColor, CityFurnitureDeviceType, CityFurnitureTarget
 from traffic_control.admin.audit_log import AuditLogHistoryAdmin
 from traffic_control.mixins import EnumChoiceValueDisplayAdminMixin
 
@@ -53,23 +47,3 @@ class CityFurnitureTargetAdmin(AuditLogHistoryAdmin):
     )
     ordering = ("name_fi",)
     actions = None
-
-
-class ResponsibleEntityAdminForm(forms.ModelForm):
-    def __init__(self, *args, **kwargs):
-        super().__init__(*args, **kwargs)
-        parent_choices = ((re.id, re) for re in ResponsibleEntity.objects.exclude(id=self.instance.id))
-        self.fields["parent"].choices = (("", "---"), *parent_choices)
-
-
-@admin.register(ResponsibleEntity)
-class ResponsibleEntityAdmin(AuditLogHistoryAdmin):
-    list_display = (
-        "id",
-        "name",
-        "organization_level",
-        "parent",
-    )
-    ordering = ("organization_level",)
-    actions = None
-    form = ResponsibleEntityAdminForm
