+++ city_furniture/filters.py
@@ -2,13 +2,13 @@ from django_filters import Filter
 from django_filters.rest_framework import FilterSet
 from rest_framework.exceptions import NotFound
 
-from city_furniture.models import FurnitureSignpostPlan, FurnitureSignpostReal, FurnitureSignpostRealOperation
-from city_furniture.models.common import (
-    CityFurnitureColor,
-    CityFurnitureDeviceType,
-    CityFurnitureTarget,
+from city_furniture.models import (
+    FurnitureSignpostPlan,
+    FurnitureSignpostReal,
+    FurnitureSignpostRealOperation,
     ResponsibleEntity,
 )
+from city_furniture.models.common import CityFurnitureColor, CityFurnitureDeviceType, CityFurnitureTarget
 from traffic_control.filters import GenericMeta, OperationalAreaFilter
