+++ city_furniture/models/__init__.py
@@ -1,4 +1,5 @@
 # flake8: noqa
+from .common import CityFurnitureColor, CityFurnitureDeviceType, CityFurnitureTarget
 from .furniture_signpost import (
     CityFurnitureColor,
     CityFurnitureDeviceType,
@@ -8,5 +9,5 @@ from .furniture_signpost import (
     FurnitureSignpostReal,
     FurnitureSignpostRealFile,
     FurnitureSignpostRealOperation,
-    ResponsibleEntity,
 )
+from .responsible_entity import ResponsibleEntity
