+++ city_furniture/models/furniture_signpost.py
@@ -9,12 +9,7 @@ from django.utils.translation import gettext_lazy as _
 from enumfields import Enum, EnumField, EnumIntegerField
 
 from city_furniture.enums import CityFurnitureDeviceTypeTargetModel
-from city_furniture.models.common import (
-    CityFurnitureColor,
-    CityFurnitureDeviceType,
-    CityFurnitureTarget,
-    ResponsibleEntity,
-)
+from city_furniture.models.common import CityFurnitureColor, CityFurnitureDeviceType, CityFurnitureTarget
 from traffic_control.enums import Condition, InstallationStatus, Lifecycle
 from traffic_control.mixins.models import (
     AbstractFileModel,
@@ -59,7 +54,7 @@ class FurnitureAbstractSignpost(SourceControlModel, SoftDeleteModel, UserControl
         on_delete=models.PROTECT,
     )
     responsible_entity = models.ForeignKey(
-        ResponsibleEntity,
+        "ResponsibleEntity",
         verbose_name=_("Responsible entity"),
         blank=True,
         null=True,
