+++ city_furniture/resources/furniture_signpost.py
@@ -1,13 +1,8 @@
 from import_export.fields import Field
 from import_export.widgets import ForeignKeyWidget
 
-from city_furniture.models import FurnitureSignpostPlan, FurnitureSignpostReal
-from city_furniture.models.common import (
-    CityFurnitureColor,
-    CityFurnitureDeviceType,
-    CityFurnitureTarget,
-    ResponsibleEntity,
-)
+from city_furniture.models import FurnitureSignpostPlan, FurnitureSignpostReal, ResponsibleEntity
+from city_furniture.models.common import CityFurnitureColor, CityFurnitureDeviceType, CityFurnitureTarget
 from city_furniture.resources.common import (
     GenericDeviceBaseResource,
     ResourceEnumIntegerField,
