+++ city_furniture/serializers/responsible_entity.py
@@ -1,7 +1,7 @@
 from enumfields.drf import EnumSupportSerializerMixin
 from rest_framework import serializers
 
-from city_furniture.models.common import ResponsibleEntity
+from city_furniture.models import ResponsibleEntity
 
 
 class ResponsibleEntitySerializer(EnumSupportSerializerMixin, serializers.ModelSerializer):
